renamed aw2.go -> main.go for clarity, also added some lipgloss test
